Enabled the CUDA kernel with num_mc_tasks = 256 by default. Removed boost timer. Wrapped CUDA function by the checkCudaErrors helper. Initialized seed in the Ctor of kernel.

+++ src/receptor.hpp
@@ -19,7 +19,7 @@ public:
 	const array<float, 3> corner1; ///< Box boundary corner with largest values of all the 3 dimensions.
 	const float granularity; ///< 1D size of grids.
 	const float granularity_inverse; ///< 1 / granularity.
-	array<size_t, 3> num_probes; ///< Number of probes.
+	array<int, 3> num_probes; ///< Number of probes.
 	size_t num_probes_product; ///< Product of num_probes[0,1,2]
 	vector<vector<size_t>> p_offset;
 	vector<vector<float>> maps;
